feat(hydroflow_lang): Added state_by operator.
